Added Color and spawn animation to paddles
I added different colors to each paddle to help the player identify which interactable items are either helpful or harmful to them.

There will be two variants of the triangle pickup: each color coded to help associate them with a specific player.

The ball's color will also take on the color of the last paddle it was in contact with. This influences how the ball interacts with each triangle pickup.